+++ tests/unit/core/test_views.py
@@ -5,10 +5,12 @@ import json
 import pytest
 from directory_api_client import api_client
 
+from django.db.utils import DataError
 from django.urls import reverse
 
 from core import helpers, models, serializers
 from tests.helpers import create_response
+from tests.unit.core.factories import ListPageFactory, DetailPageFactory
 from tests.unit.learn.factories import LessonPageFactory, TopicPageFactory
 
 
@@ -121,12 +123,13 @@ def test_dashboard_page_lesson_progress(
     client.force_login(user)
 
     # given the user has read some lessons
-    topic_one = TopicPageFactory(parent=domestic_homepage, slug='topic-one', record_read_progress=True)
-    topic_two = TopicPageFactory(parent=domestic_homepage, slug='topic-two', record_read_progress=True)
-    lesson_one = LessonPageFactory(parent=topic_one, slug='lesson-one')
-    lesson_two = LessonPageFactory(parent=topic_one, slug='lesson-two')
-    LessonPageFactory(parent=topic_one, slug='lesson-three',)
-    LessonPageFactory(parent=topic_one, slug='lesson-four')
+    topic_one = ListPageFactory(parent=domestic_homepage, slug='topic-one', record_read_progress=True)
+    topic_two = ListPageFactory(parent=domestic_homepage, slug='topic-two', record_read_progress=True)
+    lesson_one = DetailPageFactory(parent=topic_one, slug='lesson-one')
+    lesson_two = DetailPageFactory(parent=topic_one, slug='lesson-two')
+    DetailPageFactory(parent=topic_one, slug='lesson-three',)
+    DetailPageFactory(parent=topic_one, slug='lesson-four')
+    DetailPageFactory(parent=topic_two, slug='lesson-one-topic-two')
     models.PageView.objects.create(
         page=lesson_one,
         list_page=topic_one,
@@ -150,7 +153,31 @@ def test_dashboard_page_lesson_progress(
     assert response.context_data['list_pages'][0].read_progress == 50
     assert response.context_data['list_pages'][1] == topic_two
     assert response.context_data['list_pages'][1].read_count == 0
-    assert response.context_data['list_pages'][1].read_progress is None
+    assert response.context_data['list_pages'][1].read_progress == 0
+
+
+@pytest.mark.django_db
+@mock.patch.object(api_client.personalisation, 'events_by_location_list')
+@mock.patch.object(api_client.personalisation, 'export_opportunities_by_relevance_list')
+def test_dashboard_page_lesson_division_by_zero(
+        mock_events_by_location_list,
+        mock_export_opportunities_by_relevance_list,
+        mock_get_company_profile,
+        client,
+        user,
+        domestic_homepage,
+        domestic_site
+):
+    mock_events_by_location_list.return_value = create_response(json_body={'results': []})
+    mock_export_opportunities_by_relevance_list.return_value = create_response(json_body={'results': []})
+    client.force_login(user)
+
+    # given a lesson listing page without any lesson in it
+    ListPageFactory(parent=domestic_homepage, slug='test-topic-one', record_read_progress=True)
+
+    # when the dashboard is visited a division by zero should be raised
+    with pytest.raises(DataError, match='division by zero'):
+        client.get(reverse('core:dashboard'))
 
 
 @pytest.mark.django_db